Added some more tests

+++ packages/hfs/test/create/create.test.ts
@@ -22,11 +22,11 @@ var deleteFolderRecursive = function(path) {
         fs.rmdirSync(path);
     }
 };
-const save = (files:any) => {
-    if (fs.existsSync(dir)) {
+const save = (files:any, clean = true) => {
+    if (fs.existsSync(dir) && clean) {
         deleteFolderRecursive(dir)
     }
-    fs.mkdirSync(dir);
+    !fs.existsSync(dir) && fs.mkdirSync(dir);
     files.forEach((f: any) => fs.writeFileSync(npath.join(dir, f.path), f.content));
 }
 const run = (path:string) => {
@@ -102,5 +102,148 @@ export const c = 1 + b;
             expect(msg).toBe("abc are 1, 1, 2");
             done();
         };
+    });
+
+    test(`Scenario: Source file import unknown file, error "dependency is not found" should be emit
+    Given c.js
+    When calling create with stream "^-c-$"
+      | c | { path: "c.js", content: c.js, required: true } |
+    Then create return a stream "!'dependency ./b.js is not found'"`, (done) => {
+        expect.assertions(1);
+        const input: hfs.File[] = [
+            {
+                path: './c.js',
+                content: files["./c.js"]
+            },
+        ]
+        const fs = new HFS();
+        fs.create(from(input))
+            .subscribe(
+                (f: any) => {},
+                (e)=>{
+                    expect(e).toBe('dependency ./b.js is not found');
+                    done();
+                },
+                ()=>{
+
+                }
+            );
+    });
+    test(`Scenario: File that isn't flag with required and not import form required file should not output
+    Given c.js
+    When calling create with stream "^-c-$"
+      | c | { path: "b.js", content: b.js } |
+    Then create return a stream "^-$"`, (done) => {
+        throw "not implemented";
+        expect.assertions(1);
+        const input: hfs.File[] = [
+            {
+                path: './c.js',
+                content: files["./c.js"]
+            },
+        ]
+        const fs = new HFS();
+        fs.create(from(input))
+            .subscribe(
+                (f: any) => {
+                    expect(true).toBe("observable shouldn't emit")
+                },
+                (e)=>{
+
+
+                },
+                ()=>{
+                    expect(true).toBe(true);
+                    done();
+                }
+            );
+    });
+    test(`  Scenario: Deploying modified application on top of the old version, both version should work
+    Given files a.js, b.js and c.js
+    And Calling create with stream "^-a-b-c-$"
+      | a | { path: "a.js", content: a.js, required: true } |
+      | b | { path: "b.js", content: b.js } |
+      | c | { path: "c.js", content: c.js } |
+    And saving the output of create
+    When File a.js -> const a = 1 is changed to 2
+    And running create and saving out put on top of previous output
+    Then Running the outputted version of old a.js it should alert "abc are 1, 1, 2"
+    And  Running the outputted version of new a.js it should alert "abc are 2, 1, 2"`, (done) => {
+        expect.assertions(2);
+        const fs = new HFS();
+
+        const input1: hfs.File[] = [
+            {
+                path: './a.js',
+                content: files["./a.js"]
+            },
+            {
+                path: './b.js',
+                content: files["./b.js"]
+            },
+            {
+                path: './c.js',
+                content: files["./c.js"]
+            },
+        ]
+        const output1: any = [];
+
+        fs.create(from(input1))
+            .subscribe(
+                (f: any) => output1.push(f),
+                (e)=>{ throw e},
+                ()=>{
+                    save(output1);
+
+                }
+            );
+        const input2: hfs.File[] = [
+            {
+                path: './a.js',
+                content: `import { b } from "./b";
+import { c } from "./c";
+
+const a = 2;
+alert(\`abc are \${a}, \${b}, \${c}\`);`
+            },
+            {
+                path: './b.js',
+                content: files["./b.js"]
+            },
+            {
+                path: './c.js',
+                content: files["./c.js"]
+            },
+        ]
+        const output2: any = [];
+
+        fs.create(from(input2))
+            .subscribe(
+                (f: any) => output2.push(f),
+                (e)=>{ throw e},
+                ()=>{
+                    save(output2, false);
+
+                    (global as any).alert = (msg) => {
+                        expect(msg).toBe("abc are 1, 1, 2");
+                        (global as any).alert = (msg) => {
+                            expect(msg).toBe("abc are 2, 1, 2");
+                            done();
+                        };
+
+                        output2.forEach((f:any) => f.source[0] === "./a.js" && run("./" + npath.join('./tmp', f.path)) );
+                    };
+
+                    output1.forEach((f:any) => f.source[0] === "./a.js" && run("./" + npath.join('./tmp', f.path)) );
+
+
+
+
+                }
+            );
+
+
+
+
     });
 });
